chore: :coffin: move dead code to discontinued folder
