+++ lib/view/home/discontinued_pages/challenges.dart
