+++ lib/view/home/discontinued_pages/principal.dart
@@ -1,8 +1,8 @@
 import 'package:bookclub/common/style_manager.dart';
-import 'package:bookclub/view/home/challenges.dart';
-import 'package:bookclub/view/home/favorite_books.dart';
+import 'package:bookclub/view/home/discontinued_pages/challenges.dart';
+import 'package:bookclub/view/home/discontinued_pages/favorite_books.dart';
 import 'package:bookclub/view/home/gender.dart';
-import 'package:bookclub/view/home/list_books.dart';
+import 'package:bookclub/view/home/discontinued_pages/list_books.dart';
 import 'package:flutter/material.dart';
 
 class PrincipalPage extends StatefulWidget {
