+++ lib/view/home/discontinued_pages/profile_detail.dart
